requirements.txt
cachelib==0.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-Login==0.5.0
Flask-SocketIO==4.3.0
Flask-SQLAlchemy==2.4.3
gevent==20.6.2
gevent-websocket==0.10.1
greenlet==0.4.16
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==7.1.2
psycopg2==2.8.5
python-dotenv==0.13.0
python-engineio==3.13.0
python-socketio==4.6.0
requests==2.23.0
six==1.15.0
SQLAlchemy==1.3.17
urllib3==1.25.9
Werkzeug==1.0.1
zope.event==4.4
zope.interface==5.1.0
